Heuristic testing issues repaired

+++ src/app/camp-related/camp-list/camp-list.component.ts
@@ -2,6 +2,9 @@ import {Component, OnInit} from '@angular/core';
 import {CampService} from '../../service/camp.service';
 import {Camp} from '../../model/camp';
 import {ChildService} from '../../service/child.service';
+import {MenuService} from '../../service/menu.service';
+import {MatDialog} from '@angular/material';
+import {ConfirmDialogComponent} from '../../structure/confirm-dialog/confirm-dialog.component';
 
 @Component({
     selector: 'app-camp-list',
@@ -13,15 +16,22 @@ export class CampListComponent implements OnInit {
     camps: Camp[];
 
     constructor(private campService: CampService,
-                private childService: ChildService) {
+                private childService: ChildService,
+                public dialog: MatDialog) {
     }
 
     ngOnInit() {
         this.campService.getUserCamps().subscribe(camps => this.camps = camps);
     }
 
-    removeCamp(id: number) {
-        this.campService.removeCamp(id);
+    removeCamp(id: number, event) {
+        const dialogRef = this.dialog.open(ConfirmDialogComponent, {data: 'Opravdu chcete smazat tento tábor?'});
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                this.campService.removeCamp(id);
+            }
+        });
+        event.stopPropagation();
     }
 
     getChildrenCount(id: number) {
